+++ src/loading/manifest_map.rs
@@ -24,7 +24,18 @@ impl ManifestMap
         self.map.remove(key)
     }
 
+    /// Gets a file reference from a scene file reference.
+    ///
+    /// Returns `None` if the requested file is [`SceneFile::ManifestKey`] and lookup failed.
+    pub(crate) fn get(&self, file: &SceneFile) -> Option<SceneFile>
+    {
+        let SceneFile::ManifestKey(key) = file else { return Some(file.clone()) };
+        self.map.get(key).cloned()
+    }
+
     /// Swaps a manifest key for a file reference.
+    ///
+    /// Logs an error if looking up a manifest key fails.
     pub(crate) fn swap_for_file(&self, maybe_key: &mut SceneFile)
     {
         let SceneFile::ManifestKey(key) = maybe_key else { return };